+++ src/main/java/net/jasper/mod/gui/option/CommandsToExcludeOption.java
@@ -31,7 +31,7 @@ public class CommandsToExcludeOption extends Screen {
 
     // Should no longer use a singleton when used in general
     public static String FILE_NAME = "ignored_commands.txt";
-    public static String FILE_PATH = Path.of(MinecraftClient.getInstance().runDirectory.getAbsolutePath(), FILE_NAME).toString();
+    public static String FILE_PATH = Path.of(PlayerAutomaClient.PLAYERAUTOMA_FOLDER_PATH, FILE_NAME).toString();
 
     public static final List<String> userCommandsToIgnore = new ArrayList<>();